don't assume the unlocked device will appear after udev_settle()
Call wait_for() instead to ensure that the unlocked device appears.
Note that udev_settle() might not be called by wait_for() if the
unlocked device already exists.  This means that the recently added
wait for the LVM logical volume device is now necessary even if the LV
device events are guaranteed to be put on the udev queue before
'vgchange -a y' returns.