wait for the LVM LV device to appear before reading its fs type
By itself, this change is probably not necessary.  udev_settle() is
called after unlocking the device, which should guarantee that
'vgchange -a y' has finished running before this point.  The logical
volume device event triggered by 'vgchange -a y' is almost certainly
guaranteed to be on the udev queue before vgchange returns, so the
udev_settle() after unlocking the device should guarantee that the LV
device is present before this point.

However, this check is being added for two reasons:

  * If 'vgchange -a y' might return before the LV device event has
    been added to the udev queue then udev_settle() might return
    before the LV device has been created.  That opens up a possible
    race condition that this change fixes.

  * The next commit will skip the call to udev_settle() in certain
    circumstances, which will make this check necessary regardless of
    how 'vgchange' behaves.

Note that the overhead of wait_for() is miniscule if the LVM LV device
already exists by the time wait_for() is executed.